ExBanking.User.Bucket (ExBanking v0.1.0)

User bucket to store user information

Functions

add_amount(user, amount, currency)

Will increase the user money amount with the given amount and currency

add_amount!(user, amount, currency)

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

decrease_amount(user, amount, currency)

Will decrease the user money amount with the given amount and currency

decrease_amount!(user, amount, currency)

get_amount(user, currency)

Specs

get_amount(user :: String.t(), currency :: String.t()) ::
  number() | {:error, :not_found}

Returns the user money amount

get_amount!(user, currency)

registry_name(user)

start_link(name)

update_user_balance(user, new_balance, currency)