ExBanking.User.Producer (ExBanking v0.1.0)

User producer logic following the GenStage documentation. Here we can call the sync_notify function for passing events to our consumer.

Example

iex> ExBanking.User.Producer.sync_notify({:deposit, %{user: "Paulo", amount: 2.00, currency: "BRL"}})
2.0

Types

event()

Specs

event() ::
  {:deposit | :withdraw | :get_balance, ExBanking.User.Producer.EventParam.t()}

Functions

handle_call(event, from, state)

Handler for the GenStage call function.

If the pending_demand is less than zero this handler will reply with an error based on the consumer max_demand. But if this event has a priority, like a rollback in a transaction, the handler will dispatch immediately.

handle_demand(incoming_demand, arg)

Handler for the GenStage demands.

It will receive the GenStage demands and dispatch to the event handler. This will increase the demands by adding the incoming demand to the pending demand

init(atom)

Init the GenStage producer

registry_exists?(user)

Specs

registry_exists?(user :: String.t()) :: boolean()

Checks if the user producer exists

registry_name(user)

Name registry for User Producer

start_link(name)

Starts the user producer.

sync_notify(event, timeout \\ 5000)

Specs

sync_notify(event(), :infinity | non_neg_integer()) ::
  {:ok, any()} | {:error, atom(), atom()}

Sends an event and returns only after the event is dispatched.